# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FeedbackRequest {
  /** @minLength 1 */
  question: string;
  transcript: string;
}

export type FeedbackFocusItemExample = {
  youSaid: string;
  tryInstead: string;
};

export interface FeedbackFocusItem {
  title: string;
  description: string;
  example: FeedbackFocusItemExample;
}

export type FeedbackResponseWhatWentWell = {
  summary: string;
  tags: string[];
};

export type FeedbackResponseTrySayingItThisWay = {
  suggestion: string;
  why: string;
};

export interface FeedbackResponse {
  whatWentWell: FeedbackResponseWhatWentWell;
  focusOn: FeedbackFocusItem[];
  trySayingItThisWay: FeedbackResponseTrySayingItThisWay;
}

export interface ErrorResponse {
  error: string;
}

export interface UserProfile {
  name: string | null;
}

export interface UpdateProfileRequest {
  /** @minLength 1 */
  name: string;
}

export interface PracticeSessionAnswer {
  question: string;
  transcript: string;
  feedback: FeedbackResponse;
}

export interface PracticeSession {
  id: number;
  /** @minItems 1 */
  answers: PracticeSessionAnswer[];
  completedAt: string;
}

export interface CreateSessionRequest {
  /** @minItems 1 */
  answers: PracticeSessionAnswer[];
}

export interface SessionListResponse {
  sessions: PracticeSession[];
}