# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: feedback
    description: AI feedback operations
  - name: profile
    description: Authenticated user profile operations
  - name: sessions
    description: Practice session history operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /feedback:
    post:
      operationId: generateFeedback
      tags: [feedback]
      summary: Generate AI feedback for a practice response
      description: Sends a question and transcript to Gemini and returns structured coaching feedback.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
      responses:
        "200":
          description: Feedback generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: AI feedback is not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me:
    get:
      operationId: getMe
      tags: [profile]
      summary: Get the current authenticated user's profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateMe
      tags: [profile]
      summary: Set the current authenticated user's preferred name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sessions:
    get:
      operationId: listSessions
      tags: [sessions]
      summary: List the current authenticated user's completed practice sessions
      description: Ordered most-recent first.
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createSession
      tags: [sessions]
      summary: Record a completed practice session for the current authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "200":
          description: Session recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeSession"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    FeedbackRequest:
      type: object
      required:
        - question
        - transcript
      properties:
        question:
          type: string
          minLength: 1
        transcript:
          type: string
    FeedbackFocusItem:
      type: object
      required:
        - title
        - description
        - example
      properties:
        title:
          type: string
        description:
          type: string
        example:
          type: object
          required:
            - youSaid
            - tryInstead
          properties:
            youSaid:
              type: string
            tryInstead:
              type: string
    FeedbackResponse:
      type: object
      required:
        - whatWentWell
        - focusOn
        - trySayingItThisWay
      properties:
        whatWentWell:
          type: object
          required:
            - summary
            - tags
          properties:
            summary:
              type: string
            tags:
              type: array
              items:
                type: string
        focusOn:
          type: array
          items:
            $ref: "#/components/schemas/FeedbackFocusItem"
        trySayingItThisWay:
          type: object
          required:
            - suggestion
            - why
          properties:
            suggestion:
              type: string
            why:
              type: string
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    UserProfile:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          nullable: true
    UpdateProfileRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
    PracticeSessionAnswer:
      type: object
      required:
        - question
        - transcript
        - feedback
      properties:
        question:
          type: string
        transcript:
          type: string
        feedback:
          $ref: "#/components/schemas/FeedbackResponse"
    PracticeSession:
      type: object
      required:
        - id
        - answers
        - completedAt
      properties:
        id:
          type: number
        answers:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/PracticeSessionAnswer"
        completedAt:
          type: string
          format: date-time
    CreateSessionRequest:
      type: object
      required:
        - answers
      properties:
        answers:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/PracticeSessionAnswer"
    SessionListResponse:
      type: object
      required:
        - sessions
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/PracticeSession"