# lib/api-zod/src/generated/types/createSessionRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PracticeSessionAnswer } from './practiceSessionAnswer';

export interface CreateSessionRequest {
  /** @minItems 1 */
  answers: PracticeSessionAnswer[];
}