# lib/api-zod/src/generated/types/practiceSession.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PracticeSessionAnswer } from './practiceSessionAnswer';

export interface PracticeSession {
  id: number;
  /** @minItems 1 */
  answers: PracticeSessionAnswer[];
  completedAt: Date;
}