# lib/api-zod/src/generated/types/practiceSessionAnswer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackResponse } from './feedbackResponse';

export interface PracticeSessionAnswer {
  question: string;
  transcript: string;
  feedback: FeedbackResponse;
}