# lib/api-zod/src/generated/types/sessionListResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PracticeSession } from './practiceSession';

export interface SessionListResponse {
  sessions: PracticeSession[];
}